fix(plugin): add plugin for head; clean mixin
- add $speedkitHead for apply head in custom definition
