+++ lib/module.js
@@ -73,6 +73,12 @@ async function addBuildTemplates (scope, options) {
     }
   });
 
+  scope.addPlugin({
+    src: path.resolve(__dirname, 'templates/plugin/head.js'),
+    fileName: pkg.name + '/plugin/head.js',
+    options: {}
+  });
+
   scope.addTemplate({
     src: path.resolve(__dirname, 'templates/entry.js'),
     fileName: pkg.name + '/entry.js',
