+++ lib/templates/plugin.js
@@ -46,4 +46,3 @@ export default (context, inject) => {
   }
 });
 
-Vue.config.optionMergeStrategies.head = Vue.config.optionMergeStrategies.data;
